requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-openai
langchain-mongodb
pymongo[srv]
motor
pyyaml
python-dotenv
